+++ ggml/src/ggml-hexagon/htp/hex-dma.c
@@ -1,4 +1,4 @@
-#include "htp-dma.h"
+#include "hex-dma.h"
 
 #include <stdbool.h>
 #include <stdlib.h>
