+++ ggml/src/ggml-hexagon/htp/htp-ops.h
@@ -4,10 +4,11 @@
 #include "htp-ctx.h"
 #include "htp-msg.h"
 #include "worker-pool.h"
-#include "ops-utils.h"
 
 #include <assert.h>
 #include <stdint.h>
+
+#include <hex-fastdiv.h>
 
 // ggml-common.h must be included prior to this header
 
@@ -81,7 +82,7 @@ struct htp_ops_context {
     struct fastdiv_values cpy_rshp_div_n0;       // fastdiv values for ne00
     struct fastdiv_values cpy_rshp_div_n1n0;     // fastdiv values for ne00*ne01
     struct fastdiv_values cpy_rshp_div_n2n1n0;   // fastdiv values for ne00*ne01*ne02
-    
+
     uint32_t flags;
 };
